Move steps up or down in the sequence editor. Fixed delay bug (#6)
There are now two buttons on the seuqence editor tab that let steps get moved up or down the list of steps. Can be used to insert step midway by adding it tot he bottom then moving it up. The bug fix was that steps without the dealy selected, could have a value input in the delay field that would delay the sequence even if that input was hidden. Now the delay only occurs on "Delay" steps, which is how it should have been