Log file updates
Added writing buffer so that we can limit the max file size of the log files, defaulted to 10Mb. I also added a limit on the process_samples array that is used for plotting the data to 1M samples. At high sampling rates this will limit what is shown on the plots but should keep things from bogging down as the code runs during a test, which is bad